+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0
+
+- Includes new API with `MapboxItemBuilder`, `MapboxItem`, and `MapboxAutoTransformItem`
+- `MapboxItemBuilder` exposes the `screenPosition` `zoom`, `bearing`, and `tilt` in a `builder`-closure to reactively build widgets according to the exposed attributes.
+
 ## 0.0.1
 
 - Initial version.